move zqd/api to api
