+++ api/jsonpipescanner.go
