+++ cmd/zapi/cmd/index/find.go
@@ -3,12 +3,12 @@ package idx
 import (
 	"flag"
 
+	"github.com/brimsec/zq/api"
 	"github.com/brimsec/zq/archive"
 	"github.com/brimsec/zq/cli/outputflags"
 	"github.com/brimsec/zq/cmd/zapi/cmd"
 	"github.com/brimsec/zq/emitter"
 	"github.com/brimsec/zq/zbuf"
-	"github.com/brimsec/zq/zqd/api"
 	"github.com/mccanne/charm"
 )
