+++ cmd/zapi/cmd/info/info.go
@@ -8,10 +8,10 @@ import (
 	"reflect"
 	"text/tabwriter"
 
+	"github.com/brimsec/zq/api"
 	"github.com/brimsec/zq/cmd/zapi/cmd"
 	"github.com/brimsec/zq/cmd/zapi/format"
 	"github.com/brimsec/zq/pkg/nano"
-	"github.com/brimsec/zq/zqd/api"
 	"github.com/mccanne/charm"
 )
