+++ cmd/zapi/cmd/newsubspace/newsubspace.go
@@ -5,8 +5,8 @@ import (
 	"flag"
 	"fmt"
 
+	"github.com/brimsec/zq/api"
 	"github.com/brimsec/zq/cmd/zapi/cmd"
-	"github.com/brimsec/zq/zqd/api"
 	"github.com/brimsec/zq/zqd/storage"
 	"github.com/mccanne/charm"
 )
