+++ cmd/zapi/cmd/post/pcap.go
@@ -9,10 +9,10 @@ import (
 	"sync/atomic"
 	"time"
 
+	"github.com/brimsec/zq/api"
 	"github.com/brimsec/zq/cmd/zapi/cmd"
 	"github.com/brimsec/zq/cmd/zapi/format"
 	"github.com/brimsec/zq/pkg/display"
-	"github.com/brimsec/zq/zqd/api"
 	"github.com/mccanne/charm"
 )
