+++ driver/multisource.go
@@ -4,13 +4,13 @@ import (
 	"context"
 	"io"
 
+	"github.com/brimsec/zq/api"
 	"github.com/brimsec/zq/ast"
 	"github.com/brimsec/zq/field"
 	"github.com/brimsec/zq/filter"
 	"github.com/brimsec/zq/pkg/nano"
 	"github.com/brimsec/zq/scanner"
 	"github.com/brimsec/zq/zng/resolver"
-	"github.com/brimsec/zq/zqd/api"
 )
 
 // A MultiSource is a set of one or more ZNG record sources, which could be
