+++ zqd/search/indexsearch.go
@@ -3,10 +3,10 @@ package search
 import (
 	"context"
 
+	"github.com/brimsec/zq/api"
 	"github.com/brimsec/zq/archive"
 	"github.com/brimsec/zq/zbuf"
 	"github.com/brimsec/zq/zng/resolver"
-	"github.com/brimsec/zq/zqd/api"
 )
 
 type IndexSearcher interface {
