+++ zqd/space/filespace.go
@@ -6,8 +6,8 @@ import (
 	"os"
 	"sync"
 
+	"github.com/brimsec/zq/api"
 	"github.com/brimsec/zq/pkg/iosrc"
-	"github.com/brimsec/zq/zqd/api"
 	"github.com/brimsec/zq/zqe"
 )
